refactor: apply review-helper guidelines to pagination utils
- Split buildKeysetPaginationFilter.ts into one-util-per-file with
  .util.ts suffix per naming convention
- Use existing combineFilters from twenty-shared instead of inline copy
- Extract isOrderByDirection, resolveOrderByFields, extractOrderByFieldNames,
  reverseOrderBy into separate files
