+++ packages/twenty-front/src/modules/object-record/record-show/utils/isOrderByDirection.util.ts
@@ -1,0 +1,9 @@
+const ORDER_DIRECTIONS = new Set([
+  'AscNullsFirst',
+  'AscNullsLast',
+  'DescNullsFirst',
+  'DescNullsLast',
+]);
+
+export const isOrderByDirection = (value: unknown): value is string =>
+  typeof value === 'string' && ORDER_DIRECTIONS.has(value);
